+++ samples/Demo/Beef.Demo.Test/PersonTest.cs
@@ -185,7 +185,7 @@ namespace Beef.Demo.Test
                 .ExpectStatusCode(HttpStatusCode.NotModified)
                 .RunOverride(() => new PersonAgent(new DemoWebApiAgentArgs(AgentTester.GetHttpClient(), beforeRequestAsync: async r =>
                 {
-                    r.Headers.IfNoneMatch.Add(new System.Net.Http.Headers.EntityTagHeaderValue("\"" + p.ETag + "\""));
+                    r.Headers.IfNoneMatch.Add(new System.Net.Http.Headers.EntityTagHeaderValue(p.ETag));
                     await Task.CompletedTask;
                 })).GetAsync(3.ToGuid()));
         }