+++ src/Beef.Data.Database/CHANGELOG.md
@@ -4,6 +4,7 @@ Represents the **NuGet** versions.
 
 ## v4.2.5
 - *Enhancement:* Upgraded the `Microsoft.Data.SqlClient` to the latest version. This may break existing tooling and testing; to correct add `;TrustServerCertificate=true` to the connection string.
+- *Fixed:* The `DatabaseRowVersionConverter` has been fixed to ensure that the `ETag` must be quoted to be compliant. See https://datatracker.ietf.org/doc/html/rfc7232#page-9 where specification states `entity-tag = [ weak ] opaque-tag`, and `opaque-tag = DQUOTE *etagc DQUOTE`; therefore, should be formatted as: `"value"`.
 
 ## v4.2.4
 - *Fixed:* Issue [157](https://github.com/Avanade/Beef/issues/157) fixed. Renamed `DatabasePropertyMapper.DbType` method to be `SetDbType` to be consistent with the other related _set_ methods.